requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
faiss-cpu
tiktoken
python-dotenv
pymupdf
